// src/erp-ui/projects/api/src/lib/model/booking-category-dto.ts
/**
 * Fachschaften-ERP.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TaxArea } from './tax-area';


export interface BookingCategoryDto { 
    id: string;
    name: string;
    taxArea: TaxArea;
}