// src/erp-ui/projects/api/src/lib/model/cash-book-dto.ts
/**
 * Fachschaften-ERP.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CashBookDto { 
    parentName?: string | null;
    isClosed: boolean;
    closedAt?: string | null;
    balance: number;
    id: string;
    name: string;
    parentId?: string | null;
}