// src/erp-ui/projects/api/src/lib/model/cash-income-request.ts
/**
 * Fachschaften-ERP.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TaxArea } from './tax-area';


export interface CashIncomeRequest { 
    date: string;
    cashAmount: number;
    tallyListAmount: number;
    sumUpAmount: number;
    taxArea: TaxArea;
    cashBookId: string | null;
    categoryId: string | null;
    description: string | null;
}